# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Access-Nex
  description: >
    Self-hosted OAuth2 / OpenID Connect provider. This document covers the
    JSON surface a client application (including a separately-hosted
    frontend — see docs/FRONTEND.md) can build against: the standard
    OIDC/OAuth2 endpoints, the /api/v1 account API, and the /api/admin
    management API. The server-rendered HTML pages (/login, /portal, /admin,
    /device) are not part of this contract — they exist for zero-setup
    browser use and can be replaced entirely by a custom frontend calling
    these APIs.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Default local install
security:
  - sessionCookie: []
paths:
...
  /api/v1/login:
    post:
      summary: Password login (step 1)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { username: { type: string }, password: { type: string } }
      responses:
        "200":
          description: Signed in, or a 2FA challenge
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties: { status: { type: string, enum: [signed_in] } }
                  - type: object
                    properties:
                      status: { type: string, enum: [2fa_required] }
                      token: { type: string, description: "pass to /api/v1/login/totp or /login/webauthn/*" }
                      methods: { type: array, items: { type: string, enum: [totp, webauthn] } }
        "401": { description: Invalid credentials }
        "429": { description: Rate limited }
...
  /login/webauthn/begin:
    post:
      summary: Password login (step 2, security key) — get assertion options
      security: []
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { token: { type: string } } }
      responses:
        "200":
          description: WebAuthn PublicKeyCredentialRequestOptions
          content:
            application/json:
              schema:
                type: object
                properties: { waToken: { type: string }, publicKey: { type: object } }
  /login/webauthn/finish:
    post:
      summary: Password login (step 2, security key) — verify assertion
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                waToken: { type: string }
                credential: { type: object, description: "PublicKeyCredential.toJSON() from navigator.credentials.get()" }
      responses:
        "200": { description: Signed in; sets the session cookie }
        "400": { description: Verification failed }
...
  /portal/webauthn/register/finish:
    post:
      summary: Register a new passkey/security key (step 2)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                name: { type: string, description: "user-facing label" }
                credential: { type: object, description: "PublicKeyCredential.toJSON() from navigator.credentials.create()" }
      responses:
        "201": { description: Registered }

  # ── /api/admin: management API (admin session required) ────────────────────
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oidc_sid
      description: >
        Set by /login, /api/v1/login(+totp), or /login/webauthn/finish.
        Mutating /api/v1 and /api/admin requests additionally require the
        X-Access-Nex-Api / X-Access-Nex-Admin header respectively.
    clientBasic:
      type: http
      scheme: basic
      description: OAuth2 client_id:client_secret (confidential clients); public clients omit auth entirely
    bearerToken:
      type: http
      scheme: bearer
    dpopToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'For DPoP-bound tokens: "Authorization: DPoP <token>" plus a DPoP proof header'